test: move bulk repayment log tests to their own doctype
test_bulk_repayment_logs and test_bulk_payments_for_multiple_disbursements
assert on Bulk Repayment Log status, not on Loan Repayment behaviour, so
they belong with the Bulk Repayment Log doctype rather than in the already
large loan repayment test file.

Move both into bulk_repayment_log/test_bulk_repayment_log.py (with the same
shared class setup). No test logic changes; both still pass.
